+++ menu.py
@@ -50,7 +50,7 @@ def move():
         if move == "forward":
             print("\nYou moved 500m forward")
             print("You've come across Immortal Man")
-            inventory()
+            inventory2()
         elif move == "backward":
             print("\nYou moved 500m backward")
             print("You've found extra flashlight batteries")
@@ -70,7 +70,7 @@ def move():
             continue
 
 
-def inventory():
+def inventory2():
     """Creates a function for the inventory of the character.
     """
     # Creates a dictionary storing both tool and heal items.
@@ -135,4 +135,4 @@ def inventory():
 
 start()
 move()
-inventory()
+inventory2()