+++ changelog.md
@@ -2,8 +2,8 @@
 v0.0
 - added READ.me for important information
 - added changelog to show modifications
-- added main game file with RPG Nested Dictionaries
+- added main game file which imports functions from multiple files
 - added pseudocode file with comments
 - added menu file with functions
-- added map file for map game
+- added map file for the different maps of the game
 - added info file for info about characters, locations, and items